+++ src/unistd/closefrom_close_range_syscall.c
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2022 Guilherme Janczak <guilherme.janczak@yandex.com>
+ *
+ * Permission to use, copy, modify, and distribute this software for any
+ * purpose with or without fee is hereby granted, provided that the above
+ * copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ * ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ */
+
+#include <sys/syscall.h>
+
+#include <unistd.h>
+
+#include "closefrom.h"
+
+int
+close_range(unsigned int first, unsigned int last, int flags)
+{
+	/* Use __NR because some libcs don't have a SYS_ decl. */
+#if !defined(__NR_close_range)
+	#error "Linux < 5.9 is unsupported, update your kernel and headers."
+#endif
+	return syscall(__NR_close_range, first, last, (unsigned int)flags);
+}